Move exceptions into their own package
