+++ graylog2-server/src/main/java/org/graylog2/contentpacks/exceptions/InvalidParameterTypeException.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.contentpacks;
+package org.graylog2.contentpacks.exceptions;
 
 import org.graylog2.contentpacks.model.entities.references.ValueType;
